Set up contact artist page
Sets up the skeleton for for the "Contact the Artist" page. Adds the
show and create routes for a ContactController that will be used to
send emails to Dan.

Moves the navbar into its own partial and sets up an additional nav_only
layout that can be used on pages other than the main page.